* commit 'b0c8123f41508455bf1c58e856fee0870c404c31':
  UPDATE comments and jdoc
  FIX reporter assignation bug and contributor update tests
  ADDED tests
  OPTIMIZE update constributor service
  FIX long times to update contributor
  [DELETE]delete exceptions
  [ADD]add new tests
  [ADD]add one test
  add junits
  [ADD]add tests
  [ADD]add tests

+++ dataflow-service/src/main/java/org/eea/dataflow/service/impl/ContributorServiceImpl.java
@@ -13,6 +13,7 @@ import org.eea.interfaces.vo.contributor.ContributorVO;
 import org.eea.interfaces.vo.dataflow.DataFlowVO;
 import org.eea.interfaces.vo.dataset.DesignDatasetVO;
 import org.eea.interfaces.vo.dataset.ReportingDatasetVO;
+import org.eea.interfaces.vo.ums.ResourceAccessVO;
 import org.eea.interfaces.vo.ums.ResourceAssignationVO;
 import org.eea.interfaces.vo.ums.ResourceInfoVO;
 import org.eea.interfaces.vo.ums.UserRepresentationVO;
@@ -50,7 +51,7 @@ public class ContributorServiceImpl implements ContributorService {
 
   /** The dataflow controlle zuul. */
   @Autowired
-  private DataFlowControllerZuul dataflowControlleZuul;
+  private DataFlowControllerZuul dataflowControllerZuul;
 
   /** The user management controller zull. */
   @Autowired
@@ -74,7 +75,7 @@ public class ContributorServiceImpl implements ContributorService {
     List<ContributorVO> contributorVOList = new ArrayList<>();
 
     if (EDITOR.equals(role) || REPORTER.equals(role)) {
-      DataFlowVO dataflow = dataflowControlleZuul.findById(dataflowId);
+      DataFlowVO dataflow = dataflowControllerZuul.findById(dataflowId);
       Long referenceId = EDITOR.equals(role) ? dataflowId
           : dataflow.getReportingDatasets().stream()
               .filter(
@@ -125,7 +126,7 @@ public class ContributorServiceImpl implements ContributorService {
   @Override
   public void deleteContributor(Long dataflowId, String account, String role, Long dataProviderId)
       throws EEAException {
-    DataFlowVO dataflow = dataflowControlleZuul.findById(dataflowId);
+    DataFlowVO dataflow = dataflowControllerZuul.findById(dataflowId);
 
     ResourceGroupEnum resourceGroupEnumWrite = null;
     ResourceGroupEnum resourceGroupEnumRead = null;
@@ -183,7 +184,7 @@ public class ContributorServiceImpl implements ContributorService {
   @Override
   public void createContributor(Long dataflowId, ContributorVO contributorVO, String role)
       throws EEAException {
-    DataFlowVO dataflow = dataflowControlleZuul.findById(dataflowId);
+    DataFlowVO dataflow = dataflowControllerZuul.findById(dataflowId);
     SecurityRoleEnum securityRoleEnum = null;
     ResourceGroupEnum resourceGroupEnum = null;
     ResourceGroupEnum resourceGroupEnumDataflow = null;
@@ -215,13 +216,13 @@ public class ContributorServiceImpl implements ContributorService {
         break;
     }
     final List<ResourceAssignationVO> resourceAssignationVOList = new ArrayList<>();
+    List<ResourceInfoVO> resourceInfoVOs = new ArrayList<>();
     if (EDITOR.equals(role)) {
 
       ResourceInfoVO resourceDataflow =
           resourceManagementControllerZull.getResourceDetail(dataflowId, resourceGroupEnumDataflow);
       if (null == resourceDataflow.getName()) {
-        resourceManagementControllerZull
-            .createResource(createGroup(dataflowId, ResourceTypeEnum.DATAFLOW, securityRoleEnum));
+        resourceInfoVOs.add(createGroup(dataflowId, ResourceTypeEnum.DATAFLOW, securityRoleEnum));
       }
       resourceAssignationVOList.add(fillResourceAssignation(dataflowId, contributorVO.getAccount(),
           resourceGroupEnumDataflow));
@@ -229,7 +230,7 @@ public class ContributorServiceImpl implements ContributorService {
         ResourceInfoVO resourceDataSchema = resourceManagementControllerZull
             .getResourceDetail(designDatasetVO.getId(), resourceGroupEnum);
         if (null == resourceDataSchema.getName()) {
-          resourceManagementControllerZull.createResource(
+          resourceInfoVOs.add(
               createGroup(designDatasetVO.getId(), ResourceTypeEnum.DATA_SCHEMA, securityRoleEnum));
         }
         resourceAssignationVOList.add(fillResourceAssignation(designDatasetVO.getId(),
@@ -239,7 +240,7 @@ public class ContributorServiceImpl implements ContributorService {
       ResourceInfoVO resourceDataflow =
           resourceManagementControllerZull.getResourceDetail(dataflowId, resourceGroupEnumDataflow);
       if (null == resourceDataflow.getName()) {
-        resourceManagementControllerZull.createResource(
+        resourceInfoVOs.add(
             createGroup(dataflowId, ResourceTypeEnum.DATAFLOW, SecurityRoleEnum.REPORTER_READ));
       }
       resourceAssignationVOList.add(fillResourceAssignation(dataflowId, contributorVO.getAccount(),
@@ -249,10 +250,10 @@ public class ContributorServiceImpl implements ContributorService {
         ResourceInfoVO resourceDataSchema = resourceManagementControllerZull
             .getResourceDetail(reportingDatasetVO.getId(), resourceGroupEnum);
         if (null == resourceDataSchema.getName()) {
-          resourceManagementControllerZull.createResource(
+          resourceInfoVOs.add(
               createGroup(reportingDatasetVO.getId(), ResourceTypeEnum.DATASET, securityRoleEnum));
-          resourceManagementControllerZull.createResource(createGroup(dataflowId,
-              ResourceTypeEnum.DATA_SCHEMA, SecurityRoleEnum.REPORTER_READ));
+          resourceInfoVOs.add(createGroup(dataflowId, ResourceTypeEnum.DATA_SCHEMA,
+              SecurityRoleEnum.REPORTER_READ));
         }
         resourceAssignationVOList.add(fillResourceAssignation(reportingDatasetVO.getId(),
             contributorVO.getAccount(), resourceGroupEnum));
@@ -261,6 +262,8 @@ public class ContributorServiceImpl implements ContributorService {
       }
 
     }
+    // Resources creation
+    resourceManagementControllerZull.createResources(resourceInfoVOs);
     // we add data to contributor
     userManagementControllerZull.addContributorsToResources(resourceAssignationVOList);
   }
@@ -316,19 +319,26 @@ public class ContributorServiceImpl implements ContributorService {
 
     // we delete the contributor and after that we create it to update
     if (EDITOR.equals(role) || REPORTER.equals(role)) {
-      try {
-        deleteContributor(dataflowId, contributorVO.getAccount(), role, dataProviderId);
-      } catch (EEAException e) {
-        LOG_ERROR.error("Error deleting contributor with the account: {} in the dataflow {} ",
-            contributorVO.getAccount(), dataflowId);
-        throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e);
+      // avoid delete if it's a new contributor
+      List<ResourceAccessVO> resourceAccessVOs =
+          userManagementControllerZull.getResourcesByUserEmail(contributorVO.getAccount());
+      if (resourceAccessVOs != null && !resourceAccessVOs.isEmpty()
+          && resourceAccessVOs.stream().anyMatch(resource -> resource.getId().equals(dataflowId)
+              && resource.getResource().equals(ResourceTypeEnum.DATAFLOW))) {
+        try {
+          deleteContributor(dataflowId, contributorVO.getAccount(), role, dataProviderId);
+        } catch (EEAException e) {
+          LOG_ERROR.error("Error deleting contributor with the account: {} in the dataflow {} ",
+              contributorVO.getAccount(), dataflowId);
+          throw new EEAException(e);
+        }
       }
       try {
         createContributor(dataflowId, contributorVO, role);
       } catch (EEAException e) {
         LOG_ERROR.error("Error creating contributor with the account: {} in the dataflow {} ",
             contributorVO.getAccount(), dataflowId);
-        throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e);
+        throw new EEAException(e);
       }
     } else {
       LOG_ERROR.error(
